# perf: improve GenericByteBuilder::append_array to use SIMD for extending the offsets