--- a/arrow-array/src/builder/generic_bytes_builder.rs
+++ b/arrow-array/src/builder/generic_bytes_builder.rs
@@ -157,13 +157,15 @@ impl<T: ByteArrayType> GenericByteBuilder<T> {
             // Shifting all the offsets
             let shift: T::Offset = self.next_offset() - offsets[0];
 
-            // Creating intermediate offsets instead of pushing each offset is faster
-            // (even if we make MutableBuffer to avoid updating length on each push
-            //  and reserve the necessary capacity, it's still slower)
-            let mut intermediate = Vec::with_capacity(offsets.len() - 1);
+            use num::Zero;
 
-            for &offset in &offsets[1..] {
-                intermediate.push(offset + shift)
+            // Creating intermediate offsets instead of pushing each offsets is faster.
+            //
+            // Not using Vec::with_capacity and push as it will not use SIMD for some reason.
+            let mut intermediate = vec![T::Offset::zero(); offsets.len() - 1];
+
+            for (index, &offset) in offsets[1..].iter().enumerate() {
+                intermediate[index] = offset + shift;
             }
 
             self.offsets_builder.extend_from_slice(&intermediate);