@alamb 2025-09-19
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                                          main                                   make-append-array-more-use-simd
-----                                                          ----                                   -------------------------------
concat 1024 arrays boolean 4                                   1.01     28.6±0.05µs        ? ?/sec    1.00     28.3±0.03µs        ? ?/sec
concat 1024 arrays i32 4                                       1.02     14.5±0.03µs        ? ?/sec    1.00     14.2±0.12µs        ? ?/sec
concat 1024 arrays str 4                                       1.00     55.9±0.13µs        ? ?/sec    1.07     60.0±0.77µs        ? ?/sec
concat boolean 1024                                            1.00    400.2±0.68ns        ? ?/sec    1.09    437.1±1.02ns        ? ?/sec
concat boolean 8192 over 100 arrays                            1.00     44.2±0.04µs        ? ?/sec    1.16     51.2±0.06µs        ? ?/sec
concat boolean nulls 1024                                      1.00    719.0±0.70ns        ? ?/sec    1.06    759.5±1.08ns        ? ?/sec
concat boolean nulls 8192 over 100 arrays                      1.00     96.3±0.19µs        ? ?/sec    1.14    110.1±0.24µs        ? ?/sec
concat fixed size lists                                        1.00   812.2±33.78µs        ? ?/sec    1.00   808.4±51.64µs        ? ?/sec
concat i32 1024                                                1.03    408.8±0.88ns        ? ?/sec    1.00    397.8±1.02ns        ? ?/sec
concat i32 8192 over 100 arrays                                1.00   239.1±10.18µs        ? ?/sec    1.02    244.2±6.82µs        ? ?/sec
concat i32 nulls 1024                                          1.00    731.4±2.95ns        ? ?/sec    1.01    738.4±1.62ns        ? ?/sec
concat i32 nulls 8192 over 100 arrays                          1.00    307.5±7.52µs        ? ?/sec    1.07   328.3±21.74µs        ? ?/sec
concat str 1024                                                1.14     14.7±0.72µs        ? ?/sec    1.00     12.9±0.53µs        ? ?/sec
concat str 8192 over 100 arrays                                1.01    103.3±0.71ms        ? ?/sec    1.00    102.2±0.97ms        ? ?/sec
concat str nulls 1024                                          1.22      7.2±0.44µs        ? ?/sec    1.00      5.9±0.17µs        ? ?/sec
concat str nulls 8192 over 100 arrays                          1.02     53.5±0.42ms        ? ?/sec    1.00     52.3±0.60ms        ? ?/sec
concat str_dict 1024                                           1.06      3.0±0.01µs        ? ?/sec    1.00      2.8±0.01µs        ? ?/sec
concat str_dict_sparse 1024                                    1.00      6.9±0.02µs        ? ?/sec    1.00      6.9±0.01µs        ? ?/sec
concat struct with int32 and dicts size=1024 count=2           1.04      7.0±0.08µs        ? ?/sec    1.00      6.7±0.04µs        ? ?/sec
concat utf8_view  max_str_len=128 null_density=0               1.00     77.3±0.39µs        ? ?/sec    1.00     77.3±0.29µs        ? ?/sec
concat utf8_view  max_str_len=128 null_density=0.2             1.00     83.7±0.23µs        ? ?/sec    1.00     83.4±0.35µs        ? ?/sec
concat utf8_view  max_str_len=20 null_density=0                1.14     87.8±0.22µs        ? ?/sec    1.00     77.1±0.21µs        ? ?/sec
concat utf8_view  max_str_len=20 null_density=0.2              1.13     94.0±0.13µs        ? ?/sec    1.00     83.6±0.25µs        ? ?/sec
concat utf8_view all_inline max_str_len=12 null_density=0      1.00     45.4±2.50µs        ? ?/sec    1.00     45.3±3.23µs        ? ?/sec
concat utf8_view all_inline max_str_len=12 null_density=0.2    1.00     53.1±2.68µs        ? ?/sec    1.00     53.0±2.99µs        ? ?/sec
```


</p>
</details>