@rluvaton 2025-09-20
Is it possible that the benchmark is not running with target cpu native?