@rluvaton 2025-09-20
I've updated the code to no longer have intermediate buffer AND use SIMD