@alamb 2025-09-22
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/scripts/gh_compare_arrow.sh) Running
Linux aal-dev 6.14.0-1014-gcp #15~24.04.1-Ubuntu SMP Fri Jul 25 23:26:08 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing make-append-array-more-use-simd (fd5a012acdb486547d765d8b835bce4787a28840) to f4840f6df1c2549ce0947305b7111edad638b445 [diff](https://github.com/apache/arrow-rs/compare/f4840f6df1c2549ce0947305b7111edad638b445..fd5a012acdb486547d765d8b835bce4787a28840)
BENCH_NAME=concatenate_kernel
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench concatenate_kernel 
BENCH_FILTER=
BENCH_BRANCH_NAME=make-append-array-more-use-simd
Results will be posted here when complete